requirements.txt
networkx>=3.0,<4.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
pytest>=7.0,<9.0
sentence-transformers>=2.2,<4.0
scikit-learn>=1.3,<2.0
numpy>=1.24,<3.0
trafilatura>=1.6,<3.0
beautifulsoup4>=4.12,<5.0
google-genai>=1.0
python-dotenv>=1.0,<2.0
praw>=7.7,<8.0
pyyaml>=6.0,<7.0
questionary>=2.0,<3.0
apscheduler>=3.10,<4.0
jinja2>=3.1,<4.0
fastapi>=0.100,<1.0
uvicorn>=0.24,<1.0
